Add support for __no_BUILD__ constructor argument
When Moo or Moose (or other frameworks) subclass a Class::Tiny class,
they want to be able to delegate construction but control the BUILD
calls themselves.  The "__no_BUILD__" argument is an agreed convention
that indicates that the constructor must not run BUILD, as the caller
is taking responsibility for doing so later.